Add license component + license logos
