+++ src/assets/scss/components/detail-overview.scss
@@ -127,5 +127,10 @@
 .cc-license-img {
 height: 15px;
 width: 15px;
-margin-left: 10px;
+margin-right: 5px;
+}
+
+.license-name {
+  padding-bottom: 5px;
+  margin: 0;
 }
